Move around schema into logic groups
